prodemo/requirement.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cors-headers==3.2.1
django-extra-fields==2.0.3
django-filter==2.2.0
django-rest-framework-redocs==0.0.8
django-rest-swagger==2.2.0
django-username-email==2.2.4
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
olefile==0.46
openapi-codec==1.3.2
packaging==20.1
Pillow==4.2.1
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.1
urllib3==1.25.8
zipp==3.0.0
